fix(consensus): ensure boundary certs are gossiped for non-committee members (#593)
This patch follows https://github.com/commonwarexyz/monorepo/pull/2022
by inverting the relationship between receiving votes and sending out
finalization certificates for boundary heights (the final heights of an
epoch):

1. the epoch manager now listens for votes in newer epochs
2. if it finds out about it being on an older epoch this way, it
requests the finalization certificate for the boundary height of its old
epoch from the voter
3. the voter sends this certificate to the running simplex engine
backing the old epoch of the node
4. this allows the node to move forward to the end of the epoch and
transition to the new epoch.

This process is repeated until the node catches up.