feat(stablecoin-exchange): enforce tick spacing (#1084)
Ref tempoxyz/docs#316

This PR introduces `TICK_SPACING` and enforces that new orders placed
after the allegretto hardfork adhere to the specified tick spacing. Note
that fill logic still iterates through each tick in the tick bitmap, and
this should be updated in the future for more a efficient search over
existing liquidity.